# docs/swagger.yaml
basePath: /v1
definitions:
  http.StatisticCategoryProductResponse:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      percentage:
        type: number
    type: object
  http.StatisticSupplierProductResponse:
    properties:
      percentage:
        type: number
      supplier_id:
        type: string
      supplier_name:
        type: string
    type: object
...
  http.errorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: leanhtuan1998hl@gmail.com
    name: tuanla
    url: https://github.com/tuanla/soa-be-test
  description: This is a simple RESTful Product Backend Service API written in Go
    using Gin web framework, MySQL database
  title: Go SOA Test (Source of Asia) API
  version: "1.0"
paths:
...
  /statistics/products-per-category:
    get:
      consumes:
      - application/json
      description: Get Statistic of category product
      produces:
      - application/json
      responses:
        "200":
          description: Statistic created
          schema:
            $ref: '#/definitions/http.StatisticCategoryProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get Statistic of category product
      tags:
      - Statistics
  /statistics/products-per-supplier:
    get:
      consumes:
      - application/json
      description: Get Statistic of supplier product
      produces:
      - application/json
      responses:
        "200":
          description: Statistic created
          schema:
            $ref: '#/definitions/http.StatisticSupplierProductResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get Statistic of supplier product
      tags:
      - Statistics
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"